+++ src/methods/class.js
@@ -1,0 +1,12 @@
+module.exports = {
+  addClass: function (element, className){
+    if(element) {
+      element.classList.add(className);
+    }
+  },
+  removeClass: function(element, className){
+    if(element) {
+      element.classList.remove(className);
+    }
+  }
+}
