+++ src/methods/eventListener.js
@@ -6,13 +6,14 @@ function isNextPageAvailable(page, pageSize, filteredReviews) {
   return page < Math.floor(filteredReviews.length / pageSize);
 }
 
-module.exports = {
-  showMoreReviews: function (pageNumber, filteredReviews) {
+function showMoreReviews(pageNumber, filteredReviews) {
     if (isNextPageAvailable(pageNumber, constant.PAGE_SIZE, filteredReviews)) {
       pageNumber++;
-      renderReviews.renderReviews(pageNumber, false, filteredReviews);
+      renderReviews(pageNumber, false, filteredReviews);
     } else {
       pageElement.moreReviewsBtn.classList.add('invisible');
     }
+    return pageNumber;
   }
-}
+
+module.exports = showMoreReviews;
