+++ src/methods/template.js
@@ -1,9 +1,9 @@
-module.exports = {
-  getTemplateClone: function(templateElement, className){
+function getTemplateClone(templateElement, className){
     if ('content' in templateElement) {
       return templateElement.content.querySelector('.' + className);
     } else {
       return templateElement.querySelector('.' + className);
     }
   }
-}
+
+module.exports = getTemplateClone;
