+++ src/values/constants.js
@@ -1,4 +1,5 @@
 module.exports = {
 	PAGE_SIZE: 3,
-	LOAD_TIMEOUT: 10000
+	LOAD_TIMEOUT: 10000,
+	REVIEWS_LOAD_URL: '//o0.github.io/assets/json/reviews.json'
 }
